requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
pathspec==0.9.0
pyTelegramBotAPI==4.4.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
